+++ CHANGELOG.md
@@ -9,6 +9,8 @@ Versioning].
 
 ## [Unreleased]
 
+## [0.2.1] - 2026-03-09
+
 ### Added
 
 - Headless browser mode via `Doco::builder().headless(true)`. Defaults to
@@ -41,6 +43,7 @@ Versioning].
 
 Initial release of the `doco` and `doco-derive` crates
 
+[0.2.1]: https://github.com/otterbuild/doco/releases/tag/v0.2.1
 [0.2.0]: https://github.com/otterbuild/doco/releases/tag/v0.2.0
 [0.1.0]: https://github.com/otterbuild/doco/releases/tag/v0.1.0
 [keep a changelog]: https://keepachangelog.com/en/1.0.0/